pyproject.toml
[project]
name = "my-first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.17",
    "ollama>=0.6.2",
    "openai>=2.33.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
]
